+++ src/__tests__/parsers/babel_typescript_parser.test.js
@@ -8,4 +8,4 @@
 import {parseTs} from '../../parsers/babel_parser';
 import {parserTests} from '../../../fixtures/parser_tests';
 
-parserTests(parseTs);
+parserTests(parseTs, true);